WSD/gRPC_demo/pyproject.toml
[tool.poetry]
name = "grpc-demo"
version = "0.1.0"
description = ""
authors = ["XoJIoDuJIHuK <XoJIoDuJIHuK@yandex.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.68.0"
grpcio-tools = "^1.68.0"
jupyter = "^1.1.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
